Figuras

Description

Figuras geométricas (esferas, planos, triángulos, rectángulos, cilindros, conos y BVH), sus intersecciones con rayos y la carga de mallas .obj.

Documentation

data Point2D

Tipo de dato básico, se usa para representar las coordenadas U V de texturas.

Constructors

Point2D

Fields

uP :: Float

vP :: Float

data Img

Textura opcional de una figura: sin imagen (Nada) o con una imagen RGB (Iñagen).

Constructors

Nada

Iñagen (Image PixelRGB8)

data Camara

Tipo compuesto, este representa la base de la camara y su posición tridimensional sobre dicha base.

Constructors

Camara Point3D Base

data Esfera

Tipo compuesto, representa una esfera, tiene un punto central, la longitud del radio, el color, las propiedades del material y su indice de reflexión.

Constructors

Esfera

Fields

centEs :: Point3D

radEs :: Float

rgbEs :: RGB

trEs :: (Float, Float, Float)

reflEs :: Float

idEs :: Int

imgEs :: Img

data Plano

Tipo compuesto, representa un plano, tiene un punto central, la dirección normal al plano, el color, las propiedades del material y su indice de reflexión.

Constructors

Plano

Fields

centPl :: Point3D

normPl :: Direction

rgbPl :: RGB

trPl :: (Float, Float, Float)

reflPl :: Float

idPl :: Int

imgPl :: Img

data Triangulo

Tipo compuesto, representa un triángulo, tiene sus 3 vertices, el color, las propiedades del material y su indice de reflexión.

Constructors

Triangulo

Fields

p0Tr :: Point3D

p1Tr :: Point3D

p2Tr :: Point3D

uv0Tr :: Point2D

uv1Tr :: Point2D

uv2Tr :: Point2D

rgbTr :: RGB

trTr :: (Float, Float, Float)

reflTr :: Float

idTr :: Int

imgTr :: Img

data Cilindro

Tipo compuesto, representa un cilindro, tiene un punto central, la longitud del radio, el color, las propiedades del material y su indice de reflexión.

Constructors

Cilindro Point3D Direction Float RGB (Float, Float, Float) Float Int

data Cono

Tipo compuesto, representa un cono, tiene un punto central, la longitud del radio, el color, las propiedades del material y su indice de reflexión.

Constructors

Cono

Fields

centCo :: Point3D

altCo :: Float

radCo :: Float

rgbCo :: RGB

trCo :: (Float, Float, Float)

reflCo :: Float

idCo :: Int

imgCo :: Img

data Rectangulo

Tipo compuesto, representa un rectángulo, tiene un punto central, las direcciones normal y tangente, longitud y anchura, el color, las propiedades del material y su indice de reflexión.

Constructors

Rectangulo

Fields

centRe :: Point3D

normRe :: Direction

tngRe :: Direction

altRe :: Float

ancRe :: Float

rgbRe :: RGB

trRe :: (Float, Float, Float)

reflRe :: Float

idRe :: Int

imgRe :: Img

data AABB

Tipo compuesto, representa una hitbox 3D de forma rectangular, tiene dos puntos que representan los vertices de cada extremo.

Constructors

AABB

Fields

p0AB :: Point3D

p1AB :: Point3D

data BVH

Tipo compuesto, representa una BVH, es un tipo especial dado que es recursivo, contiene un AABB que actua como Hitbox, subinstancias de si mismo y una lista de triángulos.

Constructors

BVH

Fields

aabb :: AABB

bvhs :: [BVH]

triangulos :: [Triangulo]

idBvh :: Int

data TrianglePos

Tipo auxiliar, representa la posición de un triángulo, solo tiene sus 3 vértices.

Constructors

TrianglePos

Fields

v1 :: Int

v2 :: Int

v3 :: Int

data Shape

Tipo especial, sería lo equivalente a una clase virtual, esta nos permite interactuar de forma transparente con su contenido sin necesidad de saber la clase concreta que contiene.

Constructors

Sphere Esfera

Plane Plano

Triangle Triangulo

Cylinder Cilindro

Rectangle Rectangulo

Acelerator BVH

Cone Cono

Null

data Obj

Tipo compuesto, contiene todas las propiedades obtenidas al colisionar un rayo con un objeto, la distancia del impacto, el color, la dirección incidente, el punto de colisión, la dirección normal de ese punto para el objeto y las propiedades internas del mismo(coeficientes).

Constructors

Obj

Fields

mindObj :: Float

rgbObj :: RGB

w0Obj :: Direction

colObj :: Point3D

normObj :: Direction

trObj :: (Float, Float, Float)

reflObj :: Float

idObj :: Int

shObj :: Shape

imgObj :: Img

getUV :: Shape -> Point3D -> (Float, Float)

Función auxiliar, para dada una SHAPE y un punto3d de la misma, devolver en coordenadas UV la posición de dicho punto.(Se emplea para las texturas)

calculateBoundingBox :: [Triangulo] -> AABB

Función auxiliar, dada una lista de triángulos, calcula su hitbox y la devuelve.

splitTriangles :: Int -> [Triangulo] -> [[Triangulo]]

Función auxiliar, dado un entero y una lista de triángulos, genera una lista de listas de triángulos, cada sublista tiene tantos triángulos como indique el entero.

buildBVH :: Int -> [Triangulo] -> BVH

Función básica, dada una lista de triángulos genera un BVH a partir de esta.

rayIntersectsAABB :: Ray -> AABB -> Bool

Función auxiliar, dado un rayo y una Hitbox comprueba si estos colisionan o no, devuelve un booleano para indicarlo.

closestIntersection :: Ray -> [Triangulo] -> (Float, Triangulo)

Función auxiliar, dado un rayo y una lista de triángulos, intersecta este con todos y devuelve la intersección más cercana.

addFig :: Shape -> [Shape] -> [Shape]

Función auxiliar, dada una figura individual, añade esta a una lista de figuras.

addFigMult :: [Shape] -> [Shape] -> [Shape]

Función básica, junta 2 listas de figuras.

encenderShapes :: [Shape] -> [Shape]

Función básica, convierte las figuras en luces de área.

encenderShape :: Shape -> Shape

Función auxiliar, convierte una figura a luz de área.

parametricShapeCollision :: Set Shape -> [Ray] -> [Set Obj]

Función básica, dada una figuro y una lista de rayos, devuelve la lista de colisiones de cada uno de los rayos con la figura.

oneCollision :: Shape -> Ray -> Obj

Función auxiliar, dada una figura y un rayo traza la colisión.

getImg :: Obj -> Image PixelRGB8

Función auxiliar, devuelve la textura de un objeto intersectado. Lanza un error si no tiene.

modShape :: Shape -> Image PixelRGB8 -> Shape

Función auxiliar, asigna una textura a una figura (cilindros y BVH se devuelven sin cambios).

modImg :: [Shape] -> Image PixelRGB8 -> Int -> [Shape]

Función auxiliar, asigna una textura a la figura de la lista con el id indicado.

getShapeID :: Shape -> Int

Función auxiliar, devuelve el id interno de las figuras.

ray1TriangleIntersection :: Point3D -> Direction -> Point3D -> Point3D -> Point3D -> Maybe (Float, Point3D)

Función auxiliar, dado un rayo(descompuesto) y un triángulo(descompuesto) cálcula la intersección.

triangleToTriangulo :: RGB -> (Float, Float, Float) -> Float -> Int -> Img -> ([Point3D], TrianglePos, [Point2D], TrianglePos) -> Triangulo

Función auxiliar, convierte de triangle a triangulo.

convertToCustomFormat :: RGB -> (Float, Float, Float) -> Float -> Int -> Img -> ([Point3D], [TrianglePos], [Point2D], [TrianglePos]) -> [Triangulo]

Función básica, convierte los tríangulos y vértices cargados al formato deseado(color,propiedades).

parseTriangle :: String -> Maybe [TrianglePos]

Función auxiliar, dada una línea del .obj parsea el triángulo que esta contiene.

parsePoint3D :: String -> Maybe Point3D

Función auxiliar, parsea una línea del .obj a punto 3D.

parsePoint2D :: String -> Maybe Point2D

Función auxiliar, parsea una línea del .obj a punto 3D.

loadObjFile :: FilePath -> IO ([Point2D], [Point3D], [[TrianglePos]])

Función básica, extrae los vértices y triángulos de un fichero .obj.

vertexToPoint3D :: Point3D -> Point3D

Función auxiliar, convierte de punto3D a punto3D (convierte contenido a floats).

findMinPositive :: Float -> Float -> Float

Función auxiliar, devuelve el mínimo(este debe ser positivo).