Files

Description

Lectura y escritura de ficheros: imágenes PPM y BMP, y serialización binaria.

Documentation

writeObject :: Binary a => FilePath -> DList a -> IO ()

Función básica, dado un objeto que instancie la clase Binary, escribe este en formato binario a un archivo en disco.

readObject :: Binary a => FilePath -> IO a

Función básica, dado un fichero binario, recupera el objeto que haya almacenado en este y lo devuelve.

leerPPM :: FilePath -> IO ([RGB], (Float, Float, Float, Float))

Función auxiliar, para leer un archivo .ppm y almacenar los píxeles en una lista.

findMaxPPM :: [ByteString] -> Float

Función auxiliar, para extraer el MAX dentro de un ppm.

findSizePPM :: [ByteString] -> (Float, Float)

Función auxiliar, extrae el tamaño de un ppm.

parsePixels :: [ByteString] -> [RGB]

Función auxiliar, para analizar una línea de píxeles y convertirla en una lista de RGB.

parsePixels' :: [RGB] -> String

Función auxiliar, convierte una lista de RGBs en un string de pixeles.

rgbToString :: RGB -> String

Función auxiliar, convierte un RGB a string.

writeBMP :: FilePath -> Int -> Int -> ByteString -> IO ()

Función auxiliar, escribe un 32-bit BMP file

writePPM :: FilePath -> Int -> Int -> String -> IO ()

Función principal, escibre en un archivo PPM la información de los pixels en formato P3.

intTo4Bytes :: Int -> ByteString

Función auxiliar, convierte un entero a un ByteString de 4 Bytes.

intTo2Bytes :: Int -> ByteString

Función auxiliar, convierte un entero a un ByteString de 2 Bytes.

pixels2BMP :: [RGB] -> ByteString

Función auxiliar, convierte una lista ordenada de RGBs a BytseString.