PathTracer

Description

Render mediante path tracing: luz directa e indirecta.

Documentation

pathTracer :: Float -> [Luz] -> Set Shape -> Int -> Obj -> StdGen -> RGB

Función principal del path tracer, calcula toda la suma de la luz indirecta.

luzDirecta :: [Luz] -> Set Shape -> Obj -> RGB

Función principal que calcula la luz directa de una escena.

luzMono :: Obj -> Luz -> Set Shape -> RGB

Función auxiliar que calcula la luz directa de una escena solo con una luz.

luzIndirecta :: Obj -> [Luz] -> Set Shape -> StdGen -> RGB

Función que calcula la luz indirecta de una escena.

luzArea :: Set Shape -> Int -> Obj -> StdGen -> RGB

Función principal que calcula la luz de una escena con una figura como luz de Area.

luzAreaRec :: Set Shape -> Obj -> StdGen -> RGB

Función que calcula la luz indirecta de una escena con una figura como luz de Area de manera recursiva.