PhotonMap

Description

Render mediante photon mapping: lanzamiento de fotones y estimación de densidad.

Documentation

createPhoton :: Float -> DList Foton -> Int -> Int -> Set Shape -> [Luz] -> StdGen -> Int -> DList Foton

Función básica, genera una lista de fotones lanzados desde diferentes luces.

selescLightSource :: [Luz] -> Int -> Int -> StdGen -> (Ray, Luz)

Función auxiliar que selecciona la luz de la que se va a lanzar el fotón.

traceRay :: Point3D -> Float -> RGB -> DList Foton -> Set Shape -> Int -> StdGen -> Obj -> DList Foton

Función auxiliar que calcula y almacena el recorrido de un fotón por la escena.

estDensPhoton :: [Foton] -> Obj -> Set Shape -> Float -> RGB

Función auxiliar que estima la densidad de fotones en un punto.

photonMap :: KdTree Float Foton -> [Luz] -> Float -> Set Shape -> Obj -> RGB

Función principal que calcula el color de un punto a partir de un kdt de fotones.

kdToRGB :: KdTree Float Foton -> Float -> Set Shape -> Obj -> RGB

Función auxiliar que calcula el color de un punto a partir de un kdt de fotones.