Tone_map

Description

Operadores de tone mapping: clamp, ecualización y corrección gamma.

Documentation

clamp :: Float -> [RGB] -> [RGB]

Función auxiliar, dado un valor máximo y una lista de RGBs, aplica la operación de clampeo con dicho valor sobre estos.

ecualization :: Float -> [RGB] -> [RGB]

Función auxiliar, dado un valor y una lista de RGBs, aplica la operación de ecualización con dicho valor sobre estos.

ecualization' :: Float -> [RGB] -> [RGB]

Función auxiliar, dado un valor y una lista de RGBs, revierte la operación de ecualización con dicho valor sobre estos.

elevateRGBPoints :: Float -> [RGB] -> [RGB]

Función auxiliar, dado un valor y una lista de RGBs, eleva estos a 1/valor.

gammaFunc :: Float -> Float -> [RGB] -> [RGB]

Función auxiliar, dado dos valores y una lista de RGBs, aplica la función gamma con dichos valores sobre estos.

gammaFunc' :: Float -> Float -> [RGB] -> [RGB]

Función auxiliar, dado dos valores y una lista de RGBs, revierte la aplicación de la función gamma con dichos valores sobre estos.